Che bato/preferences (#55)
## Preferences

### public files added
- cardSleeve1.png image of a default card sleeve

### GameBoard/page.tsx
- Added preference toggle handler
- Added Preference element.

### GameboardTypes.ts
- added preferenceToggle function to OpponentCardTryProps interface
since the settings button is in this component

### OpponentCardTry.tsx
- Added preferenceToggle on click function to settings button.

### Preferences.tsx
- Preference overlay that holds the entire content of the preferences.

### Preferences.types.ts
- Added preferenceTypes that were needed for subcomponents

### Tabs
- Added CardSleeveTab component which displays a grid of <CardSleeve>
elements.
- CurrentGameTab component with custom concede and revert state buttons.
- GameOptionsTab component where we have custom preferenceOptions ->
these are custom checkboxes since the default ones were not the same as
in the design.
- KeyboardShortcutTab component is where we set keybinds it contains the
keyboardLayout component and a grid of functions we can add click events
too.

### subComponents
- Added CardSleeve component which accepts the image and source name of
a cardsleeve. Each cardSleeves container is a button so we can easily
add selection to each one.
- KeyboardLayout component is a custom keyboardLayout which I tried to
mimick as closely to the design as possible. Each key is highlightable
so we can easily map keybinds and display them.
- PreferenceButton these are custom styled buttons seen in the
preference figma design. It accepts one parameter variant which is
either standard or concede.
- PreferenceCheckbox extends the class checkbox with custom styled
checkboxes since the default ones were not close to the design
- PreferenceOption is the row of a option in the GameOptionsTab it
accepts the name of the option like 'Mute' and description of the option
like ''Remove all in-game sounds."
- VerticalTabs new custom component to render vertical tabs in the
Preference view. It accepts an array of tab names which it then renders.
Currently support vertical tabs are: currentGame, keyboardShortcuts,
cardSleeves, gameOptions

---------

Co-authored-by: Dan Bastin <dan.bastin@gmail.com>