Addressing the Base Damage and Deployed Leader styling.
Added a PNG red background how I saw other backgrounds being created. Tested both single and double digit numbers of base damage. Also adjusted the background for the leader card in the center display when the leader is deployed.

I could not locate a way to determine if the leader's or base's epic action had been used to add the styling for that in as well.