Addressing initial comments.
Updated the image from a png to an svg, and added styles to give it the nice drop-shadow filter behind the svg. Updated the deployed leader border to match the styling of the unit arena borders. Confirmed the player names remain in the leader area when a leader is deployed.