Addressing the Base Damage and Deployed Leader styling. (#67)
Added a PNG red background how I saw other backgrounds being created.
Tested both single and double digit numbers of base damage. Also
adjusted the background for the leader card in the center display when
the leader is deployed.

Double check my logic for determining if a Leader is deployed or not - I
had an additional check for it on new line 116 that checked if it was in
game, the variant was leader, and the card type was not leader. I
believe that checking this way should also cover pilots from set 4
(which might be a leader upgrade?).

I could not locate a way to determine if the leader's or base's epic
action had been used to add the styling for that exhaust icon in as
well.

---------

Co-authored-by: Dan Bastin <dan.bastin@gmail.com>