Disable done button for 0.5s on resource window (#281)
We were having an issue where multiple players were reporting that the
"done" button at the resource stage pops up so fast that that they were
accidentally double-tapping it after hitting "pass" to end the round.

Added a timer that disables the "done" button for 0.5s whenever a new
resource phase happens to prevent this.

Co-authored-by: Addison Mayberry <admayber@microsoft.com>