Board and UnitBoard layout changes (#296)
The scope of this project was to make scaling changes to the UnitBoard
and the Border Edges around the board. The goal was to enforce some more
strict tiling / sizing of the game cards given the various resolutions.
We want to prevent units crossing the center line *for almost all cases*
but it will still cross the center on small resolution and very large
unit counts (8+). I also created a debug class - BreakPointOverlay so
you can enable it and get visual feedback on which 'responsive
threshold' you're operating with

A few additional things had to be fixed after this rescaling:
- I added a 3rd segment to the border outlines that can dynamically
scale (it's just a horizontal line that fills inbetween the left and
right side). I scaled down the pixel width of the small curve, and let
the horizontal line fill the rest.
- I fixed the upgrade icon / text on smaller images. You can see in the
first image that it was rendering more box then it needed. This is still
not perfect and should probably get redesigned, but it works.

Here is a good before and after, using the iPadAir resolution:

iPadAir before:

![image](https://github.com/user-attachments/assets/e1914758-f404-48fa-a893-c1cc2f4799ad)

iPadAir after:

![image](https://github.com/user-attachments/assets/ddf50b30-8a71-4d02-b676-eb00d0e5766a)

iPhone12 before:

![image](https://github.com/user-attachments/assets/947af449-6ff4-440a-bdf0-0bb8009d0282)

iPhone12 after:  (Still needs work, but closer!)

![image](https://github.com/user-attachments/assets/bb8ad19a-ee59-4732-84b5-f4545578673a)

Example of the border separating on wider monitors (this is now fixed by
the 3rd segment that fills the gap)

![image](https://github.com/user-attachments/assets/a5cc51a4-42c5-4f89-ab91-431d463a3ad6)

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>