Better handling for dropped lobby state (#306)
Looks like the FE is potentially not transitioning correctly to the
countdown screen if there is a disconnect right when the connection
happens, possibly due to the lobby state getting dropped and then not
listened for again.

Added some logic to check for it more aggressively, not 100% sure if
this solves the issue but seems to help in some tests I tried.

Co-authored-by: Addison Mayberry <admayber@microsoft.com>