Render promptState choices as tray buttons if player has selected cards (#545)
In order to increase readability this PR renders prompt choices as tray
buttons instead of in the popup modal.

Currently this is also limited to if there are exactly 2 choices, as
that's the current tray button UI limit.

<img width="1120" height="1252" alt="image"
src="https://github.com/user-attachments/assets/422396af-6438-4eb0-bae8-12106b60f25e"
/>

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>